requirements.txt
pytest==7.3.1
selene==2.0.0rc3post1
allure-pytest==2.13.2
requests
colorama==0.4.6
pypdf2~=2.11.2
openpyxl~=3.0.1
python-dotenv==0.20.0